Audit board hold metadata and support fractional depths (#313)
* Audit and correct Contact board holds

* Complete sourced hold depth ranges

* Support fractional hold measurements

* Validate hold measurements before range construction

* Allow fractional Workbench hold measurements

* Reject zero Workbench hold depths

* Clear stale hold depth validation

* Fix fractional fallback depth resolution

+++ HangTenTests/BoardTargetSubstitutionTests.swift
@@ -7,8 +7,8 @@ final class BoardTargetSubstitutionTests: XCTestCase {
         kind: HoldKind = .edge,
         feature: HoldFeature? = nil,
         fingerCapacity: Int? = nil,
-        sizeMillimeters: Int? = nil,
-        depthRangeMillimeters: ClosedRange<Int>? = nil,
+        sizeMillimeters: Double? = nil,
+        depthRangeMillimeters: ClosedRange<Double>? = nil,
         x: Double = 0
     ) -> BoardHold {
         BoardHold(
